+++ source/cppfs/source/FileHandle.cpp
@@ -220,7 +220,7 @@ FileIterator FileHandle::end() const
     return FileIterator();
 }
 
-unsigned int FileHandle::size() const
+uint64_t FileHandle::size() const
 {
     return m_backend ? m_backend->size() : 0;
 }
